m1923: config: Add pinner config.
Add default pinned files for Marlin / Sailfish.

From the documentation:
https://source.android.com/devices/tech/debug/jank_jitter#io-issues
[..]
PinnerService. Added in Android 7.0, PinnerService enables
the framework to lock some files in the page cache.
This removes the memory for use by any other process,
but if there are some files that are known a priori to be
used regularly, it can be effective to mlock those files.
[..]

bug 28251566

Change-Id: Iebdf37df9d7a5748166aed3cc21191e4d2ffa523